% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_response-class.R
\name{fda_response}
\alias{fda_response}
\alias{fda_response.formula}
\alias{fda_response.default}
\alias{is.fda_response}
\alias{print.fda_response}
\alias{summary.fda_response}
\title{A fda_response object containing a greta_array for use with function-valued response variables}
\usage{
fda_response(y, ...)

\method{fda_response}{formula}(
  formula,
  data,
  bins = NULL,
  spline_settings = list(),
  priors = list(),
  errors = "iid",
  ...
)

\method{fda_response}{default}(
  y,
  x,
  z = NULL,
  bins = NULL,
  spline_settings = list(),
  priors = list(),
  errors = "iid",
  ...
)

is.fda_response(x)

\method{print}{fda_response}(x, ...)

\method{summary}{fda_response}(object, ...)
}
\arguments{
\item{y}{a \code{matrix} or \code{data.frame} with the response variable (one row per observation)}

\item{...}{further arguments passed to or from other methods}

\item{formula}{formula describing the model to be fitted, in the format used by \link[lme4]{lmer}}

\item{data}{a named list containing the variables in \code{formula}}

\item{bins}{a vector of values at which the response variable is recorded}

\item{spline_settings}{a named list of settings to pass to the spline function (see \link[splines]{bs} for details)}

\item{priors}{a named list of settings for prior distributions (nu_local, nu_global, slab_df, scale_global, slab_scale, sigma_mean, sigma_sd, sigma_gamma)}

\item{errors}{a character denoting the type of errors in a matrix model; currently only 'iid' and 'ar1' are implemented}

\item{x}{a \code{matrix} or \code{data.frame} of predictor variables or a fda_response object for is.fda_response and print methods}

\item{z}{a \code{matrix} or \code{data.frame} of random effects variables}

\item{object}{an \code{fda_response} object}
}
\value{
An object of class \code{fda_response}, which can be used in a \link[greta]{model}
}
\description{
A \code{fda_response} object contains a function regression object
  that can be used to model a function-valued response with \link[greta]{model}
}
\examples{

library(greta.fda)

# fit an example model
fda_response <- fda_response(y ~ x1 + x2 + (1 | z1),
                             data = example_fda_data,
                             priors = list(
                               nu_local = 2,
                               nu_global = 2,
                               slab_df = 2,
                               scale_global = 2,
                               slab_scale = 2,
                               sigma_mean = 0.0,
                               sigma_sd = 5.0,
                               sigma_gamma = 5.0
                             )))

\dontrun{                 
# fit a greta model
sigma_response <- uniform(min = 0.0, max = 5.0, dim = 1)
distribution(example_fda_data$y) <- normal(fda_response$mu, sd = sigma_response)

greta_model <- with(fda_response, model(mu, alpha, beta, sigma_main, sigma_response))
}

# check if an object is a fda_response object
  
\dontrun{
is.fda_response(x)
}

# Print information about a 'fda_response' object

\dontrun{
print(x)
}

# Summarise a 'fda_response' object

\dontrun{
summary(object)
}
}
