% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta_fda-class.R
\name{greta_fda}
\alias{greta_fda}
\alias{greta_fda.formula}
\alias{greta_fda.default}
\alias{is.greta_fda}
\alias{print.greta_fda}
\alias{plot.greta_fda}
\alias{summary.greta_fda}
\alias{coef.greta_fda}
\alias{fitted.greta_fda}
\alias{predict.greta_fda}
\title{A greta_fda regression model with function-valued data}
\usage{
greta_fda(y, ...)

\method{greta_fda}{formula}(formula, data, family = "gaussian", link = NULL,
  bins = NULL, greta_settings = list(), spline_settings = list(),
  priors = list(), errors = "iid", ...)

\method{greta_fda}{default}(y, x, z = NULL, family = "gaussian",
  link = NULL, bins = NULL, greta_settings = list(),
  spline_settings = list(), priors = list(), errors = "iid", ...)

is.greta_fda(model)

\method{print}{greta_fda}(x, ...)

\method{plot}{greta_fda}(x, ...)

\method{summary}{greta_fda}(x, ...)

\method{coef}{greta_fda}(x, ...)

\method{fitted}{greta_fda}(x, ...)

\method{predict}{greta_fda}(x, ..., newdata = NULL, type = c("link",
  "response"), re.form = NULL, fun = NULL)
}
\arguments{
\item{y}{a \code{matrix} or \code{data.frame} with the response variable (one row per observation)}

\item{...}{further arguments passed to or from other methods}

\item{formula}{formula describing the model to be fitted, in the format used by \link[lme4]{lmer}}

\item{data}{a named list containing the variables in \code{formula}}

\item{family}{a GLM family passed as a \code{character}, see \link[stats]{family} (currently only gaussian (default) and poisson families are implemented)}

\item{link}{link function to be used for model family, see \link[stats]{family} for details (currently not implemented)}

\item{bins}{a vector of values at which the response variable is recorded}

\item{greta_settings}{a named list of values to pass to the inference method (see \link[greta]{inference} for details)}

\item{spline_settings}{a named list of settings to pass to the spline function (see \link[splines]{bs} for details)}

\item{priors}{a named list of prior distributions in the format of \link[greta]{distributions}}

\item{errors}{a character denoting the type of errors; currently only 'iid' is implemented}

\item{x}{a \code{matrix} or \code{data.frame} of predictor variables}

\item{z}{a \code{matrix} or \code{data.frame} of random effects variables}

\item{model}{a fitted \code{greta_fda} model}
}
\value{
An object of class \code{greta_fda}, which has associated `print`, `plot`, and `summary` methods
}
\description{
A \code{greta_fda} object contains a function regression model fitted with \code{greta_fda}
}
\examples{

library(greta.fda)

# fit an example model
model <- greta_fda(y ~ x1 + x2 + (1 | z1),
                   data = greta_fda_data,
                   family = "poisson",
                   greta_settings = list(n_samples = 100,
                                         warmup = 100,
                                         chains = 2),
                   priors = list(intercept = normal(0, 1),
                                 coefficients = normal(0, 1),
                                 random_intercepts = normal(0, 1)))
                        
\dontrun{                 
# summarise fitted model
model
summary(model)
plot(model)
}

# fit a greta_fda model with formula
  
\dontrun{
model <- greta_fda(y ~ x1 + x2 + (1 | z1),
                   data = greta_fda_data,
                   greta_settings = list(n_samples = 100,
                                         warmup = 100,
                                         chains = 2))
}

# fit a greta_fda model with separate arguments
  
\dontrun{
model <- greta_fda(y = y, x = x, z = z,
                   greta_settings = list(n_samples = 100,
                                         warmup = 100,
                                         chains = 2))
}

# check if an object is a greta_fda model
  
\dontrun{
is.greta_fda(model)
}

# Print information about a 'greta_fda' object

\dontrun{
print(x)
}

# Plot a 'greta_fda' object

\dontrun{
plot(x)
}

# Summarise a 'greta_fda' object

\dontrun{
summary(x)
}

# Extract coefficients from a 'greta_fda' object

\dontrun{
coef(x)
}

# Calculate fitted values from a 'greta_fda' object

\dontrun{
fitted(x)
}

# Predict values from a 'greta_fda' object

\dontrun{
predict(x)
}
}
