% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-datasets.R
\docType{data}
\name{example_fda_data}
\alias{example_fda_data}
\title{Example data sets for functional data analysis}
\format{A list with three elements: response matrix (y), covariates (x1, x2), and a grouping variable (z1)}
\usage{
example_fda_data
}
\description{
Example data set with a functional response variable, two covariates, and one grouping variable
}
\examples{
library(greta.fda)
names(example_fda_data)
head(example_fda_data$y)
}
\keyword{datasets}
